requirements.txt
appium-python-client
selenium
pytest
allure-pytest
requests
azure-devops>=7.1.0
python-dotenv>=0.19.0